Rename testsupport package
