+++ integration_tests/agp/src/test/java/org/robolectric/integrationtests/agp/TestActivityTest.java
@@ -4,6 +4,7 @@ import androidx.test.core.app.ActivityScenario;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.robolectric.integrationtests.agp.testsupport.TestActivity;
 
 /**
  * Test asserting that test-only activities can be declared in a dependency project's manifest as a
