+++ integration_tests/agp/testsupport/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.robolectric.integrationtests.agp">
+    package="org.robolectric.integrationtests.agp.testsupport">
   <application>
     <activity android:name="TestActivity"/>
   </application>
