+++ integration_tests/agp/testsupport/src/main/java/org/robolectric/integrationtests/agp/testsupport/TestActivity.java
@@ -1,4 +1,4 @@
-package org.robolectric.integrationtests.agp;
+package org.robolectric.integrationtests.agp.testsupport;
 
 import android.app.Activity;
